[python] shrink merge_into to MVP scope (update='*' only); fix CI
- Per Jingsong review #7, only support update='*' / insert='*'.
  Partial SET (dict) raises NotImplementedError; deferred to follow-up.
- Move WhenMatched/WhenNotMatched/_NormalizedClause/SetSpec/OnSpec to
  data_evolution_merge_transform.py to break the circular import that
  CI surfaced (data_evolution_merge_into <-> data_evolution_merge_join).
- Restore with_update_type's all-columns squash on table_update.py.
- Skip e2e tests (every case uses update='*' which still hits the
  with_update_type squash bug); re-enable after the fix PR merges.
- Replace test_blob_update_is_rejected with test_blob_columns_excluded
  to match the new silent-skip behavior.
- Doc: drop dict SET / allow_multiple_matches mentions.
