+++ paimon-python/pypaimon/write/table_update.py
@@ -109,6 +109,8 @@ class TableUpdate:
         for col in update_cols:
             if col not in self.table.field_names:
                 raise ValueError(f"Column {col} is not in table schema.")
+        if len(update_cols) == len(self.table.field_names):
+            update_cols = None
         self.update_cols = update_cols
         return self
